Add files via upload
Tic-Tac-Toe is a classic two-player game where players take turns with X or O. The objective is to be the first to get three of your marks in a row, either horizontally, vertically, or diagonally. It’s a simple yet strategic game that can end in a win, loss, or draw.